Create a workaround for a bottleneck. (#4)
Because of how the repository is abstracted, it's very difficult to get a "group by" working in the API without re-architecting the entire project. A more efficient method is going to be simply storing the last sensor reading with each sensor. This way it's easy to retrieve, and easy to consume from the API.